The initial phase of prototyping.
Blocks - block behavior class, needed to be recognized, stores the value of the integrity of the sovy (Heart). And also destroyed when it is lowered.

BlockSpawner - creates blocks by area.

Digger - makes it possible to determine that there is a block in front of us and it can be broken by right-clicking.

MouseYHeightController - The smart camera looks toward the object, and the script responds to the mouse's Y position. By changing the height with it, the camera changes the angle towards the object. If the object is up, it looks up; if the object is down, it looks down.

StorageBlocks - A simple repository of data and prefab references for blocks. You can find its implementation in the Assets/Data/New Storage Blocks.asset path.

UIBlock - Displays the amount of integrity of the block.